Migrate some func in other classes
